Rename the SAML core types and extract the controller data records (#318 step 1a)
Zero-behavior moves/renames toward the target architecture:
- Response -> SamlResponse and AuthRequest -> SamlAuthnRequest (Saml.cs), with
  every C# reference renamed; the SAML protocol string literals (the
  /samlp:Response/ XPath, the written AuthnRequest element, the SAMLResponse form
  field) and the SignatureScope.Response enum are untouched, and so is
  ControllerBase.Response.Headers.
- SamlResponse, SamlAuthnRequest and TimedAuthorizeState become internal sealed
  (default-internal, leaf types); AuthResponse stays public (it is a [FromBody]
  action parameter / JSON wire contract).
- AuthResponse and TimedAuthorizeState move out of the 1,980-line SSOController.cs
  into their own files (same namespace), shrinking the controller by 122 lines.

Behavior-preserving: public routes and the JSON/XML wire shapes are byte-identical.
Verified by the full suite (584 tests, 3x) and a 2-lens adversarial review (bypass
PASS, correctness PASS) on the diff; the earlier scoping workflow adversarially
confirmed the plan. The SsoProtocol enum at the public {mode} route boundary is
deliberately deferred (not provably zero-behavior: culture-ToLower, 500-vs-400).

Refs #318
