+++ SSO-Auth/Api/AuthResponse.cs
@@ -1,0 +1,32 @@
+namespace Jellyfin.Plugin.SSO_Auth.Api;
+
+/// <summary>
+/// The data the client should pass back to the API.
+/// </summary>
+public class AuthResponse
+{
+    /// <summary>
+    /// Gets or sets the device ID of the client.
+    /// </summary>
+    public string DeviceID { get; set; }
+
+    /// <summary>
+    /// Gets or sets the device name of the client.
+    /// </summary>
+    public string DeviceName { get; set; }
+
+    /// <summary>
+    /// Gets or sets the app name of the client.
+    /// </summary>
+    public string AppName { get; set; }
+
+    /// <summary>
+    /// Gets or sets the app version of the client.
+    /// </summary>
+    public string AppVersion { get; set; }
+
+    /// <summary>
+    /// Gets or sets the auth data of the client (for authorizing the response).
+    /// </summary>
+    public string Data { get; set; }
+}
