+++ SSO-Auth/Api/SamlResponseLoader.cs
@@ -5,8 +5,8 @@ using System.Xml;
 namespace Jellyfin.Plugin.SSO_Auth.Api;
 
 /// <summary>
-/// Safely constructs a <see cref="Response"/> from an untrusted SAML response string (#199). The
-/// <see cref="Response"/> constructor throws on malformed input — <see cref="FormatException"/> for a
+/// Safely constructs a <see cref="SamlResponse"/> from an untrusted SAML response string (#199). The
+/// <see cref="SamlResponse"/> constructor throws on malformed input — <see cref="FormatException"/> for a
 /// non-base64 body and <see cref="XmlException"/> for malformed XML or a prohibited DOCTYPE (the P2#9
 /// XXE guard). Left unhandled, those surface to an unauthenticated caller as an HTTP 500 with a stack
 /// trace; this maps them to a fail-closed <see langword="false"/> so the callback endpoints reject a
@@ -23,13 +23,13 @@ internal static class SamlResponseLoader
 
     /// <summary>
     /// Tries to parse a SAML response, returning <see langword="false"/> (rather than throwing) on the
-    /// malformed-input exceptions the <see cref="Response"/> constructor raises.
+    /// malformed-input exceptions the <see cref="SamlResponse"/> constructor raises.
     /// </summary>
     /// <param name="certificateStr">The identity provider's signing certificate as a Base64 string.</param>
     /// <param name="responseString">The untrusted SAML response (Base64).</param>
     /// <param name="response">The parsed response on success; otherwise <see langword="null"/>.</param>
     /// <returns><see langword="true"/> if the response parsed; otherwise <see langword="false"/>.</returns>
-    internal static bool TryParse(string certificateStr, string responseString, out Response response)
+    internal static bool TryParse(string certificateStr, string responseString, out SamlResponse response)
     {
         // A null or empty body is the most common malformed callback (an absent SAMLResponse form field
         // yields a null string on the unauthenticated ACS endpoint); reject it here rather than let
@@ -50,7 +50,7 @@ internal static class SamlResponseLoader
 
         try
         {
-            response = new Response(certificateStr, responseString);
+            response = new SamlResponse(certificateStr, responseString);
             return true;
         }
         catch (Exception ex) when (ex is FormatException or XmlException or CryptographicException or ArgumentException)
