+++ SSO-Auth/Api/TimedAuthorizeState.cs
@@ -1,0 +1,88 @@
+using System;
+using System.Collections.Generic;
+using Duende.IdentityModel.OidcClient;
+
+namespace Jellyfin.Plugin.SSO_Auth.Api;
+
+/// <summary>
+/// A time-stamped record of an in-flight OpenID authorize state: it ties the authorize-state token to
+/// the provider, the resolved identity (subject/username), and the login privileges, and is held until
+/// the callback redeems it or it expires.
+/// </summary>
+internal sealed class TimedAuthorizeState
+{
+    /// <summary>
+    /// Initializes a new instance of the <see cref="TimedAuthorizeState"/> class.
+    /// </summary>
+    /// <param name="state">The AuthorizeState to time.</param>
+    /// <param name="created">When this state was created.</param>
+    public TimedAuthorizeState(AuthorizeState state, DateTime created)
+    {
+        State = state;
+        Created = created;
+    }
+
+    /// <summary>
+    /// Gets or sets the Authorization State of the client.
+    /// </summary>
+    public AuthorizeState State { get; set; }
+
+    /// <summary>
+    /// Gets or sets the provider that minted this state. A state may only be consumed on the same
+    /// provider's endpoints, so it cannot be replayed against another provider's login/role gate.
+    /// </summary>
+    public string Provider { get; set; }
+
+    /// <summary>
+    /// Gets or sets when this object was created to time it out.
+    /// </summary>
+    public DateTime Created { get; set; }
+
+    /// <summary>
+    /// Gets or sets a value indicating whether the user is valid.
+    /// </summary>
+    public bool Valid { get; set; }
+
+    /// <summary>
+    /// Gets or sets the user tied to the state.
+    /// </summary>
+    public string Username { get; set; }
+
+    /// <summary>
+    /// Gets or sets the stable subject identifier (OpenID "sub") that keys the account link (#155).
+    /// Unlike <see cref="Username"/> it does not change when the identity provider renames the user.
+    /// Null for SAML, whose NameID plays the same role directly.
+    /// </summary>
+    public string Subject { get; set; }
+
+    /// <summary>
+    /// Gets or sets a value indicating whether the user is an administrator.
+    /// </summary>
+    public bool Admin { get; set; }
+
+    /// <summary>
+    /// Gets or sets a value indicating whether the state is
+    /// tied to a linking flow (instead of a login flow).
+    /// </summary>
+    public bool IsLinking { get; set; }
+
+    /// <summary>
+    /// Gets or sets the folders the user is allowed access to.
+    /// </summary>
+    public List<string> Folders { get; set; }
+
+    /// <summary>
+    /// Gets or sets a value indicating whether the user is allowed to view live TV.
+    /// </summary>
+    public bool EnableLiveTv { get; set; }
+
+    /// <summary>
+    /// Gets or sets a value indicating whether the user is allowed to manage live TV.
+    /// </summary>
+    public bool EnableLiveTvManagement { get; set; }
+
+    /// <summary>
+    /// Gets or sets the user avatar url.
+    /// </summary>
+    public string AvatarURL { get; set; }
+}
